+++ src/Command/Loader.psm1
@@ -1,3 +1,1 @@
 Import-Module ([System.IO.Path]::Combine($PSScriptRoot, 'AstralKeks.Workbench.Command.dll'))
-
-Use-WBUserspace Default